LayoutTests/http/tests/xmlhttprequest/access-control-basic-non-simple-deny-cached-expected.txt

CONSOLE MESSAGE: line 42: Request header field Content-Type is not allowed by Access-Control-Allow-Headers.
CONSOLE MESSAGE: line 42: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-basic-non-get-allow.cgi due to access control checks.
PASS: Exception thrown. Cross-domain access is not allowed in 'send'. [ A network error occurred.].
